Split streamdeck.ts into streamdeckSfx.ts and streamdeckVoice.ts
Moving the guild-resolution helpers out (previous commit) only dropped
qlty's file-total-complexity count from 59 to 52 — still flagged. The SFX
and voice-connection routes are separate concerns anyway, so split them
into their own router files; streamdeck.ts is now just composition.
