+++ src/web/routes/streamdeck.test.ts
@@ -1,5 +1,4 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import type { SfxTrigger, SfxFile } from '../../db';
 
 const API_KEY_OWNER = 'user-1';
 
@@ -11,11 +10,11 @@ vi.mock('../middleware', () => ({
 }));
 
 vi.mock('../../db', () => ({
-  findTrigger: vi.fn(),
-  findSoundFiles: vi.fn(),
   getAllSfxTriggers: vi.fn(),
   isKeyApprovedForGuild: vi.fn(),
   getApprovedGuildIdsForKey: vi.fn(),
+  findTrigger: vi.fn(),
+  findSoundFiles: vi.fn(),
 }));
 
 vi.mock('../../commands/soundSelector', () => ({
@@ -57,37 +56,7 @@ vi.mock('../../shared/logger', () => ({
 import express from 'express';
 import supertest from 'supertest';
 import router from './streamdeck';
-import { findTrigger, findSoundFiles, getAllSfxTriggers, isKeyApprovedForGuild, getApprovedGuildIdsForKey } from '../../db';
-import { pickWeightedRandom } from '../../commands/soundSelector';
-import { playFile, VoiceNotConnectedError } from '../../audio/sfxPlayer';
-import { setVoicePlaying } from '../../shared/statusStore';
-import { getAvailableVoiceChannels } from '../../discord/discordUtils';
-import { connect, disconnect } from '../../audio/audioPlayer';
-import { getDiscordClient } from '../../discord/discordBot';
-import { getActiveGuildForUser } from '../../discord/voicePresence';
-
-const TRIGGER: SfxTrigger = {
-  id: BigInt(1),
-  trigger_command: '!ding',
-  category_id: null,
-  hidden: false,
-  description: null,
-};
-
-const FILES: SfxFile[] = [
-  { id: 1, trigger_id: BigInt(1), file: 'ding.mp3', trigger_command: null, weight: 1, hidden: false, category_id: null },
-];
-
-/** Fake Discord client whose channels.fetch resolves any channel to the given guildId. */
-function makeClient(channelGuildId: string | null = 'guild-123') {
-  return {
-    channels: {
-      fetch: vi.fn(async (channelId: string) =>
-        channelGuildId ? { id: channelId, guildId: channelGuildId } : null,
-      ),
-    },
-  } as any;
-}
+import { getAllSfxTriggers, getApprovedGuildIdsForKey } from '../../db';
 
 function buildApp() {
   const app = express();
@@ -99,377 +68,17 @@ function buildApp() {
 beforeEach(() => {
   vi.clearAllMocks();
   vi.mocked(getAllSfxTriggers).mockResolvedValue([]);
-  vi.mocked(findTrigger).mockResolvedValue(null);
-  vi.mocked(findSoundFiles).mockResolvedValue([]);
-  vi.mocked(pickWeightedRandom).mockReturnValue('ding.mp3');
-  vi.mocked(getDiscordClient).mockReturnValue(makeClient());
-  vi.mocked(connect).mockResolvedValue(undefined);
-  vi.mocked(getAvailableVoiceChannels).mockResolvedValue([]);
-  vi.mocked(getActiveGuildForUser).mockReturnValue('guild-123');
-  vi.mocked(isKeyApprovedForGuild).mockResolvedValue(true);
-  vi.mocked(getApprovedGuildIdsForKey).mockResolvedValue(['guild-123']);
+  vi.mocked(getApprovedGuildIdsForKey).mockResolvedValue([]);
 });
 
-describe('POST /sfx', () => {
-  it('calls playFile with the bare filename and the presence-resolved guild', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue(FILES);
-    vi.mocked(pickWeightedRandom).mockReturnValue('only1_v2.mp3');
-
-    await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(200);
-
-    expect(vi.mocked(playFile)).toHaveBeenCalledWith('only1_v2.mp3', 'guild-123');
-    expect(vi.mocked(playFile)).not.toHaveBeenCalledWith(expect.stringContaining('/'), expect.anything());
-  });
-
-  it('returns 200 and the filename on success', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue(FILES);
-    vi.mocked(pickWeightedRandom).mockReturnValue('ding.mp3');
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true, file: 'ding.mp3' });
-    expect(vi.mocked(setVoicePlaying)).toHaveBeenCalledWith('ding.mp3', '!ding', 'streamdeck');
-  });
-
-  it('normalises the command to lowercase before lookup', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue(FILES);
-
-    await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!DING' })
-      .expect(200);
-
-    expect(vi.mocked(findTrigger)).toHaveBeenCalledWith('!ding');
-  });
-
-  it('returns 400 when command field is missing', async () => {
-    const res = await supertest(buildApp()).post('/sfx').send({}).expect(400);
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(playFile)).not.toHaveBeenCalled();
-  });
-
-  it('returns 400 when command is whitespace only', async () => {
-    const res = await supertest(buildApp()).post('/sfx').send({ command: '   ' }).expect(400);
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(playFile)).not.toHaveBeenCalled();
-  });
-
-  it('returns 503 when the key owner is not connected to voice in any guild', async () => {
-    vi.mocked(getActiveGuildForUser).mockReturnValue(null);
-
-    const res = await supertest(buildApp()).post('/sfx').send({ command: '!ding' }).expect(503);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(playFile)).not.toHaveBeenCalled();
-  });
-
-  it('returns 403 when the key is not approved for the resolved guild', async () => {
-    vi.mocked(isKeyApprovedForGuild).mockResolvedValue(false);
-
-    const res = await supertest(buildApp()).post('/sfx').send({ command: '!ding' }).expect(403);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(playFile)).not.toHaveBeenCalled();
-  });
-
-  it('returns 503 when Discord client is not ready', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(null);
-
-    const res = await supertest(buildApp()).post('/sfx').send({ command: '!ding' }).expect(503);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('returns 404 when trigger is not found', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(null);
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!unknown' })
-      .expect(404);
-
-    expect(res.body).toMatchObject({ ok: false, error: 'Unknown command' });
-  });
-
-  it('returns 404 when trigger has no sound files', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue([]);
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(404);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(playFile)).not.toHaveBeenCalled();
-  });
-
-  it('returns 503 when bot is not connected to a voice channel', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue(FILES);
-    vi.mocked(playFile).mockImplementation(() => { throw new VoiceNotConnectedError(); });
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(503);
-
-    expect(res.body).toMatchObject({ ok: false, error: 'Bot is not connected to a voice channel' });
-  });
-
-  it('returns 500 when playFile throws an unexpected error', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockResolvedValue(FILES);
-    vi.mocked(playFile).mockImplementation(() => { throw new Error('FFMPEG crashed'); });
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('returns 500 on DB error looking up trigger', async () => {
-    vi.mocked(findTrigger).mockRejectedValue(new Error('DB gone'));
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('returns 500 on DB error fetching sound files', async () => {
-    vi.mocked(findTrigger).mockResolvedValue(TRIGGER);
-    vi.mocked(findSoundFiles).mockRejectedValue(new Error('DB gone'));
-
-    const res = await supertest(buildApp())
-      .post('/sfx')
-      .send({ command: '!ding' })
-      .expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-});
-
-describe('GET /sfx', () => {
-  it('returns the list of SFX triggers', async () => {
-    const triggers = [{ triggerId: '1', triggerCommand: '!ding', description: null, hidden: false, categoryName: null, files: [] }];
-    vi.mocked(getAllSfxTriggers).mockResolvedValue(triggers as any);
-
+describe('streamdeck router composition', () => {
+  it('mounts the SFX sub-router', async () => {
     const res = await supertest(buildApp()).get('/sfx').expect(200);
-
-    expect(res.body).toEqual({ ok: true, triggers });
+    expect(res.body).toEqual({ ok: true, triggers: [] });
   });
 
-  it('returns 500 on DB error', async () => {
-    vi.mocked(getAllSfxTriggers).mockRejectedValue(new Error('DB gone'));
-
-    const res = await supertest(buildApp()).get('/sfx').expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-});
-
-describe('GET /voice/channels', () => {
-  it('returns voice channels grouped by every guild the key is approved for', async () => {
-    vi.mocked(getApprovedGuildIdsForKey).mockResolvedValue(['guild-123', '900000000000000002']);
-    const channelsA = [{ id: 'ch1', name: 'General' }];
-    const channelsB = [{ id: 'ch2', name: 'Hangout' }];
-    vi.mocked(getAvailableVoiceChannels).mockImplementation(async (guildId: string) =>
-      (guildId === 'guild-123' ? channelsA : channelsB) as any,
-    );
-
+  it('mounts the voice sub-router', async () => {
     const res = await supertest(buildApp()).get('/voice/channels').expect(200);
-
-    expect(res.body).toEqual({
-      ok: true,
-      guilds: [
-        { guildId: 'guild-123', channels: channelsA },
-        { guildId: '900000000000000002', channels: channelsB },
-      ],
-    });
-  });
-
-  it('returns 500 on error', async () => {
-    vi.mocked(getApprovedGuildIdsForKey).mockRejectedValue(new Error('DB down'));
-
-    const res = await supertest(buildApp()).get('/voice/channels').expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-});
-
-describe('POST /voice/join', () => {
-  it('returns 400 when channelId is missing', async () => {
-    const res = await supertest(buildApp()).post('/voice/join').send({}).expect(400);
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('returns 400 when channelId is not a valid Discord snowflake', async () => {
-    const res = await supertest(buildApp()).post('/voice/join').send({ channelId: 'not-a-snowflake' }).expect(400);
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('returns 503 when Discord client is not ready', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(null);
-
-    const res = await supertest(buildApp())
-      .post('/voice/join')
-      .send({ channelId: '123456789012345678' })
-      .expect(503);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('disconnects and then joins the channel, resolving the guild from the channel ID', async () => {
-    const client = makeClient('guild-123');
-    vi.mocked(getDiscordClient).mockReturnValue(client);
-
-    const res = await supertest(buildApp())
-      .post('/voice/join')
-      .send({ channelId: '123456789012345678' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true });
-    expect(vi.mocked(disconnect)).toHaveBeenCalledWith('guild-123');
-    expect(vi.mocked(connect)).toHaveBeenCalledWith(client, 'guild-123', '123456789012345678');
-    expect(vi.mocked(isKeyApprovedForGuild)).toHaveBeenCalledWith(API_KEY_OWNER, 'guild-123');
-  });
-
-  it('returns 400 when the channel does not resolve to any guild', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(makeClient(null));
-
-    const res = await supertest(buildApp())
-      .post('/voice/join')
-      .send({ channelId: '123456789012345678' })
-      .expect(400);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(connect)).not.toHaveBeenCalled();
-  });
-
-  it('returns 403 when the key is not approved for the channel\'s guild', async () => {
-    vi.mocked(isKeyApprovedForGuild).mockResolvedValue(false);
-
-    const res = await supertest(buildApp())
-      .post('/voice/join')
-      .send({ channelId: '123456789012345678' })
-      .expect(403);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(connect)).not.toHaveBeenCalled();
-  });
-
-  it('returns 500 when connect throws', async () => {
-    vi.mocked(connect).mockRejectedValue(new Error('No permission'));
-
-    const res = await supertest(buildApp())
-      .post('/voice/join')
-      .send({ channelId: '123456789012345678' })
-      .expect(500);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-});
-
-describe('POST /voice/leave', () => {
-  it('returns 400 when both channelId and guildId are missing', async () => {
-    const res = await supertest(buildApp()).post('/voice/leave').send({}).expect(400);
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('resolves the guild from channelId and disconnects', async () => {
-    const client = makeClient('guild-123');
-    vi.mocked(getDiscordClient).mockReturnValue(client);
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ channelId: '123456789012345678' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true });
-    expect(vi.mocked(disconnect)).toHaveBeenCalledWith('guild-123');
-  });
-
-  it('uses an explicit guildId directly, bypassing channel resolution entirely', async () => {
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ guildId: '900000000000000002' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true });
-    expect(vi.mocked(disconnect)).toHaveBeenCalledWith('900000000000000002');
-    expect(vi.mocked(isKeyApprovedForGuild)).toHaveBeenCalledWith(API_KEY_OWNER, '900000000000000002');
-  });
-
-  it('falls back to an explicit guildId when the channel no longer exists (e.g. deleted while connected)', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(makeClient(null));
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ channelId: '123456789012345678', guildId: '900000000000000002' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true });
-    expect(vi.mocked(disconnect)).toHaveBeenCalledWith('900000000000000002');
-  });
-
-  it('returns 400 when only channelId is given and the channel no longer exists', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(makeClient(null));
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ channelId: '123456789012345678' })
-      .expect(400);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(disconnect)).not.toHaveBeenCalled();
-  });
-
-  it('returns 403 when the key is not approved for the channel\'s guild', async () => {
-    vi.mocked(isKeyApprovedForGuild).mockResolvedValue(false);
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ channelId: '123456789012345678' })
-      .expect(403);
-
-    expect(res.body).toMatchObject({ ok: false });
-    expect(vi.mocked(disconnect)).not.toHaveBeenCalled();
-  });
-
-  it('returns 503 when Discord client is not ready and only channelId is given', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(null);
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ channelId: '123456789012345678' })
-      .expect(503);
-
-    expect(res.body).toMatchObject({ ok: false });
-  });
-
-  it('does not require the Discord client when guildId is given explicitly', async () => {
-    vi.mocked(getDiscordClient).mockReturnValue(null);
-
-    const res = await supertest(buildApp())
-      .post('/voice/leave')
-      .send({ guildId: '900000000000000002' })
-      .expect(200);
-
-    expect(res.body).toEqual({ ok: true });
-    expect(vi.mocked(disconnect)).toHaveBeenCalledWith('900000000000000002');
+    expect(res.body).toEqual({ ok: true, guilds: [] });
   });
 });
